Move screens to feature directory.
